RssParserRepository moved to :rss-parser
